keyPosition

+++ constraintlayout/core/src/main/java/androidx/constraintlayout/core/motion/utils/ArcCurveFit.java
@@ -421,7 +421,7 @@ public class ArcCurveFit extends CurveFit {
                 double pos = i / (double) (mLut.length - 1);
                 int index = Arrays.binarySearch(ourPercent, pos);
                 if (index >= 0) {
-                    mLut[i] = index / (ourPercent.length - 1);
+                    mLut[i] = index /(double) (ourPercent.length - 1);
                 } else if (index == -1) {
                     mLut[i] = 0;
                 } else {
